Refactor order price fields to use BigDecimal
Changed all order and order item price-related fields from Double to BigDecimal in DTOs, models, and integration events for improved precision in monetary calculations. Updated total calculation logic to use BigDecimal arithmetic.

+++ Backend/OrderingService/src/main/java/com/eshop/OrderingService/DTO/OrderDTO.java
@@ -2,6 +2,8 @@ package com.eshop.OrderingService.DTO;
 
 import com.eshop.OrderingService.Model.OrderStatus;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -15,5 +17,5 @@ public class OrderDTO {
     private AddressDTO address;
     private String paymentMethod;
     private List<OrderItemDTO> orderItems;
-    private Double total;
+    private BigDecimal total;
 }
